+++ crates/ruma-events/CHANGELOG.md
@@ -29,6 +29,7 @@ Improvements:
   variants in `AnyGlobalAccountDataEvent(Content)` are now prefixed with
   `Unstable`. The new `InvitePermissionConfigEventContent` struct uses the new
   format with a `default_action` field instead of `block_all`.
+- Add support for to-device event for pushing secrets, according to MSC4385.
 
 # 0.32.1